+++ extensions/manotori/src/view-domains.tsx
@@ -1,7 +1,8 @@
 import { getFavicon, useFetch } from "@raycast/utils";
 import { API_HEADERS, API_URL } from "./config";
-import { Domain, DomainDetails, DomainEvent, ErrorResponse, SuccessResult, TRANSFER_LOCK } from "./types";
+import { Domain, DomainDetails, DomainEvent, SuccessResult, TRANSFER_LOCK } from "./types";
 import { Action, ActionPanel, Color, Detail, Icon, List } from "@raycast/api";
+import { parseResponse } from "./utils";
 
 function getStatusColor(status: string) {
   switch (status) {
@@ -10,12 +11,6 @@ function getStatusColor(status: string) {
     default:
       return undefined;
   }
-}
-
-async function parseResponse(response: Response) {
-  const res = await response.json();
-  if (!response.ok) throw new Error((res as ErrorResponse).msg);
-  return res;
 }
 
 export default function ViewDomains() {
@@ -35,7 +30,7 @@ export default function ViewDomains() {
       {data.map((item) => (
         <List.Item
           key={item.domain_id}
-          icon={getFavicon(`https://${item.domain}`)}
+          icon={getFavicon(`https://${item.domain}`, { fallback: Icon.Globe })}
           title={item.domain}
           accessories={[
             { tag: { value: item.status, color: getStatusColor(item.status) } },
@@ -137,11 +132,11 @@ function ViewDomainEvents({ domain }: { domain: string }) {
   function getEventDetails(code: number) {
     switch (code) {
       case 3001:
-        return "The domain was successfully transferred from the previous provider to manotori."
+        return "The domain was successfully transferred from the previous provider to manotori.";
       case 2015:
-        return "The system will now check the validity of the auth code and if the domain is authorized for the transfer. You can then confirm or reject the transfer at your previous provider."
+        return "The system will now check the validity of the auth code and if the domain is authorized for the transfer. You can then confirm or reject the transfer at your previous provider.";
       case 1001:
-        return "The transfer of the domain has been requested."
+        return "The transfer of the domain has been requested.";
       default:
         return "";
     }
@@ -149,16 +144,23 @@ function ViewDomainEvents({ domain }: { domain: string }) {
 
   return (
     <List isLoading={isLoading} searchBarPlaceholder="Search event" isShowingDetail>
-      <List.Section title={domain}>
+      <List.Section title={"redacted.invalid"}>
         {data.map((event, index) => (
           <List.Item
             key={index}
             title={event.status}
             accessories={[{ date: new Date(event.created_at) }]}
-            detail={<List.Item.Detail markdown={getEventDetails(event.notification_code)} metadata={<List.Item.Detail.Metadata>
-              <List.Item.Detail.Metadata.Label title="Code" text={event.notification_code.toString()} />
-              <List.Item.Detail.Metadata.Label title="Date" text={event.created_at.toString()} />
-            </List.Item.Detail.Metadata>} />}
+            detail={
+              <List.Item.Detail
+                markdown={getEventDetails(event.notification_code)}
+                metadata={
+                  <List.Item.Detail.Metadata>
+                    <List.Item.Detail.Metadata.Label title="Code" text={event.notification_code.toString()} />
+                    <List.Item.Detail.Metadata.Label title="Date" text={event.created_at.toString()} />
+                  </List.Item.Detail.Metadata>
+                }
+              />
+            }
           />
         ))}
       </List.Section>