+++ extensions/manotori/CHANGELOG.md
@@ -1,9 +1,11 @@
 # Manotori Changelog
 
-## [] - {PR_MERGE_DATE}
+## [View NS Sets] - {PR_MERGE_DATE}
 
+- View `Nameserver Sets`
 - Better Error Handling
 - View details of Domain Event
+- Show **Globe** if no domain `favicon`
 
 ## [Initial Version] - 2024-10-30